created separate component for textEditor
